backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
neo4j==5.24.0
groq==0.11.0
python-dotenv==1.0.1
pydantic==2.9.2
